Move operators in a subdirectory
