+++ include/samurai/petsc.hpp
@@ -3,10 +3,12 @@
 // license that can be found in the LICENSE file.
 
 #pragma once
-#include "petsc/diffusion_FV_old.hpp"
-#include "petsc/diffusion_FV.hpp"
-#include "petsc/zero_operator_FV.hpp"
-#include "petsc/identity_FV.hpp"
+#include "petsc/fv/diffusion_FV_old.hpp"
+#include "petsc/fv/diffusion_FV.hpp"
+#include "petsc/fv/gradient_FV.hpp"
+#include "petsc/fv/divergence_FV.hpp"
+#include "petsc/fv/zero_operator_FV.hpp"
+#include "petsc/fv/identity_FV.hpp"
 #include "petsc/scheme_operators.hpp"
 #include "petsc/backward_euler.hpp"
 #include "petsc/block_backward_euler.hpp"
