+++ include/samurai/petsc/fv/diffusion_FV.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "flux_based_scheme.hpp"
+#include "../flux_based_scheme.hpp"
 
 namespace samurai 
 {
